✨ :: Seat Activity 생성, Activity 변경
